+++ spec/hashie/extensions/dash/indifferent_access_spec.rb
@@ -1,9 +1,25 @@
 require 'spec_helper'
 
 describe Hashie::Extensions::Dash::IndifferentAccess do
+  class TrashWithIndifferentAccess < Hashie::Trash
+    include Hashie::Extensions::Dash::IndifferentAccess
+    property :per_page, transform_with: lambda { |v| v.to_i }
+    property :total, from: :total_pages
+  end
+
   class DashWithIndifferentAccess < Hashie::Dash
     include Hashie::Extensions::Dash::IndifferentAccess
     property :name
+  end
+
+  context 'when included in Trash' do
+    let(:params) { { per_page: '1', total_pages: 2 } }
+    subject { TrashWithIndifferentAccess.new(params) }
+
+    it 'gets the expected behaviour' do
+      expect(subject.per_page).to eq params[:per_page].to_i
+      expect(subject.total).to eq params[:total_pages]
+    end
   end
 
   context 'when included in Dash' do